refactor(virtual_traffic_light): move to utils for unit test
